requirements.txt
runpod==1.7.7
sentence-transformers==3.3.1
open-clip-torch==2.29.0
torch==2.5.1
torchvision==0.20.1
Pillow==11.0.0
requests==2.32.3
numpy==2.2.1
